+++ example/lib/barcode_scanner_window.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:mobile_scanner/mobile_scanner.dart';
 import 'package:mobile_scanner_example/scanned_barcode_label.dart';
@@ -15,74 +14,18 @@ class BarcodeScannerWithScanWindow extends StatefulWidget {
 
 class _BarcodeScannerWithScanWindowState
     extends State<BarcodeScannerWithScanWindow> {
-  final MobileScannerController controller = MobileScannerController();
+  final MobileScannerController controller = MobileScannerController(
+    detectionSpeed: DetectionSpeed.unrestricted,
+  );
 
-  Widget _buildBarcodeOverlay() {
-    return ValueListenableBuilder(
-      valueListenable: controller,
-      builder: (context, value, child) {
-        // Not ready.
-        if (!value.isInitialized || !value.isRunning || value.error != null) {
-          return const SizedBox();
-        }
-
-        return StreamBuilder<BarcodeCapture>(
-          stream: controller.barcodes,
-          builder: (context, snapshot) {
-            final BarcodeCapture? barcodeCapture = snapshot.data;
-
-            // No barcode.
-            if (barcodeCapture == null || barcodeCapture.barcodes.isEmpty) {
-              return const SizedBox();
-            }
-
-            final scannedBarcode = barcodeCapture.barcodes.first;
-
-            // No barcode corners, or size, or no camera preview size.
-            if (value.size.isEmpty ||
-                scannedBarcode.size.isEmpty ||
-                scannedBarcode.corners.isEmpty) {
-              return const SizedBox();
-            }
-
-            return CustomPaint(
-              painter: BarcodeOverlay(
-                barcodeCorners: scannedBarcode.corners,
-                barcodeSize: scannedBarcode.size,
-                boxFit: BoxFit.contain,
-                cameraPreviewSize: value.size,
-              ),
-            );
-          },
-        );
-      },
-    );
-  }
-
-  Widget _buildScanWindow(Rect scanWindowRect) {
-    return ValueListenableBuilder(
-      valueListenable: controller,
-      builder: (context, value, child) {
-        // Not ready.
-        if (!value.isInitialized ||
-            !value.isRunning ||
-            value.error != null ||
-            value.size.isEmpty) {
-          return const SizedBox();
-        }
-
-        return CustomPaint(
-          painter: ScannerOverlay(scanWindowRect),
-        );
-      },
-    );
-  }
+  // TODO: Fix BoxFit.fill & BoxFit.fitHeight
+  final boxFit = BoxFit.contain;
 
   @override
   Widget build(BuildContext context) {
     final scanWindow = Rect.fromCenter(
-      center: MediaQuery.sizeOf(context).center(Offset.zero),
-      width: 200,
+      center: MediaQuery.sizeOf(context).center(const Offset(0, -100)),
+      width: 300,
       height: 200,
     );
 
@@ -93,22 +36,25 @@ class _BarcodeScannerWithScanWindowState
         fit: StackFit.expand,
         children: [
           MobileScanner(
-            fit: BoxFit.contain,
+            fit: boxFit,
             scanWindow: scanWindow,
             controller: controller,
-            errorBuilder: (context, error, child) {
+            errorBuilder: (context, error) {
               return ScannerErrorWidget(error: error);
             },
           ),
-          _buildBarcodeOverlay(),
-          _buildScanWindow(scanWindow),
+          BarcodeOverlay(controller: controller, boxFit: boxFit),
+          ScanWindowOverlay(
+            scanWindow: scanWindow,
+            controller: controller,
+          ),
           Align(
             alignment: Alignment.bottomCenter,
             child: Container(
               alignment: Alignment.center,
               padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
               height: 100,
-              color: Colors.black.withOpacity(0.4),
+              color: const Color.fromRGBO(0, 0, 0, 0.4),
               child: ScannedBarcodeLabel(barcodes: controller.barcodes),
             ),
           ),
@@ -121,108 +67,5 @@ class _BarcodeScannerWithScanWindowState
   Future<void> dispose() async {
     super.dispose();
     await controller.dispose();
-  }
-}
-
-class ScannerOverlay extends CustomPainter {
-  ScannerOverlay(this.scanWindow);
-
-  final Rect scanWindow;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    // we need to pass the size to the custom paint widget
-    final backgroundPath = Path()
-      ..addRect(Rect.fromLTWH(0, 0, size.width, size.height));
-    final cutoutPath = Path()..addRect(scanWindow);
-
-    final backgroundPaint = Paint()
-      ..color = Colors.black.withOpacity(0.5)
-      ..style = PaintingStyle.fill
-      ..blendMode = BlendMode.dstOver;
-
-    final backgroundWithCutout = Path.combine(
-      PathOperation.difference,
-      backgroundPath,
-      cutoutPath,
-    );
-    canvas.drawPath(backgroundWithCutout, backgroundPaint);
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) {
-    return false;
-  }
-}
-
-class BarcodeOverlay extends CustomPainter {
-  BarcodeOverlay({
-    required this.barcodeCorners,
-    required this.barcodeSize,
-    required this.boxFit,
-    required this.cameraPreviewSize,
-  });
-
-  final List<Offset> barcodeCorners;
-  final Size barcodeSize;
-  final BoxFit boxFit;
-  final Size cameraPreviewSize;
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    if (barcodeCorners.isEmpty ||
-        barcodeSize.isEmpty ||
-        cameraPreviewSize.isEmpty) {
-      return;
-    }
-
-    final adjustedSize = applyBoxFit(boxFit, cameraPreviewSize, size);
-
-    double verticalPadding = size.height - adjustedSize.destination.height;
-    double horizontalPadding = size.width - adjustedSize.destination.width;
-    if (verticalPadding > 0) {
-      verticalPadding = verticalPadding / 2;
-    } else {
-      verticalPadding = 0;
-    }
-
-    if (horizontalPadding > 0) {
-      horizontalPadding = horizontalPadding / 2;
-    } else {
-      horizontalPadding = 0;
-    }
-
-    final double ratioWidth;
-    final double ratioHeight;
-
-    if (!kIsWeb && defaultTargetPlatform == TargetPlatform.iOS) {
-      ratioWidth = barcodeSize.width / adjustedSize.destination.width;
-      ratioHeight = barcodeSize.height / adjustedSize.destination.height;
-    } else {
-      ratioWidth = cameraPreviewSize.width / adjustedSize.destination.width;
-      ratioHeight = cameraPreviewSize.height / adjustedSize.destination.height;
-    }
-
-    final List<Offset> adjustedOffset = [
-      for (final offset in barcodeCorners)
-        Offset(
-          offset.dx / ratioWidth + horizontalPadding,
-          offset.dy / ratioHeight + verticalPadding,
-        ),
-    ];
-
-    final cutoutPath = Path()..addPolygon(adjustedOffset, true);
-
-    final backgroundPaint = Paint()
-      ..color = Colors.red.withOpacity(0.3)
-      ..style = PaintingStyle.fill
-      ..blendMode = BlendMode.dstOut;
-
-    canvas.drawPath(cutoutPath, backgroundPaint);
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) {
-    return false;
   }
 }